View, Controller de Pedidos e inserção de valores ao banco

+++ src/test/java/com/fatec/lab/mineda/RestauranteApplicationTest.java
@@ -100,7 +100,7 @@ public class RestauranteApplicationTest {
 		mesa.setOpen(true);
 		mesa.setQuantidade_lugares(4);
 		
-		assertTrue((Integer)mesa.getId() != null);
+		assertTrue((Long)mesa.getId() != null);
 	}
 	
 	@Test
